+++ spec.lua
@@ -1,14 +1,13 @@
 return {
-  version = "1.0.2",
-  version_name = "artefato-astral",
+  version = "2.0.0",
+  version_name = "barian-babelico",
   install_path = {
     linux = "/usr/local/ygofab",
     windows = (os.getenv("LOCALAPPDATA") or "") .. "\\YGOFabrica"
   },
   bin_path = "/usr/local/bin",
   config_path = {
-    -- TODO: change (os.getenv("HOME") .. "/ygofab" to (os.getenv("HOME") .. "/.config/ygofab"
-    linux = (os.getenv("HOME") or "") .. "/ygofab",
+    linux = (os.getenv("HOME") or "") .. "/.config/ygofab",
     windows = (os.getenv("APPDATA") or "") .. "\\YGOFabrica"
   },
   dependencies = {